Refactor dashboard component
