--- app/Http/Livewire/Dashboard/DashboardComponent.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace App\Http\Livewire\Dashboard;
-
-use Livewire\Component;
-
-class DashboardComponent extends Component
-{
-    public function render()
-    {
-        return view('livewire.dashboard.dashboard-component');
-    }
-}
