+++ resources/views/components/dashboard-component.blade.php
@@ -1,0 +1,7 @@
+<x-app-layout>
+    @role('app.admin')
+        @livewire('dashboard.admin-dashboard-component')
+    @else
+        @livewire('dashboard.user-dashboard-component')
+    @endrole
+</x-app-layout>
