--- resources/views/livewire/dashboard/dashboard-component.blade.php
@@ -1,5 +1,0 @@
-@role('app.admin')
-    @livewire('dashboard.admin-dashboard-component')
-@else
-    @livewire('dashboard.user-dashboard-component')
-@endrole
